requirements.txt
aws-cdk.aws_ec2
aws-cdk.aws_s3
aws-cdk.aws_cloudfront
aws-cdk.aws_certificatemanager
aws-cdk.aws_elasticloadbalancingv2
aws-cdk.aws_autoscaling
cdk-spa-deploy

